+++ src/router.tsx
@@ -3,6 +3,7 @@ import { lazy } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 import { ProtectedRoute } from '@components/template/auth/ProtectedRoute';
 import { ROUTES } from '@constants/routes';
+
 const MainPage = lazy(() =>
   import('@pages/MainPage').then((module) => ({ default: module.default })),
 );
@@ -14,8 +15,10 @@ const TransactionsPage = lazy(() =>
     default: module.default,
   })),
 );
-const ApiKeysPage = lazy(() =>
-  import('@pages/ApiKeysPage').then((module) => ({ default: module.default })),
+const KeyManagementPage = lazy(() =>
+  import('@pages/KeyManagementPage').then((module) => ({
+    default: module.default,
+  })),
 );
 const SignupPage = lazy(() =>
   import('@pages/SignupPage').then((module) => ({ default: module.default })),
@@ -54,7 +57,7 @@ const routes = [
         path: ROUTES.KEY_MANAGEMENT,
         element: (
           <ProtectedRoute>
-            <ApiKeysPage />
+            <KeyManagementPage />
           </ProtectedRoute>
         ),
       },
